requirements.txt
Django==1.8.3
django-import-export==0.2.6
django-widget-tweaks==1.4.1
django-postgrespool==0.3.0
geopy==1.10.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.2
static==0.4
wsgiref==0.1.2
Pillow==2.9.0
django-allauth=0.23.0
djangorestframework
markdown
django-filter